python-service/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
requests==2.31.0
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
undetected-chromedriver==3.5.4

